requirements.txt
pytz
pandas
authlib
requests
streamlit
cryptography
python-dotenv
google-generativeai
mysql-connector-python
